NORALLY: porting v2.0 changes from #136
