+++ modules/graphman.js
@@ -2,7 +2,8 @@
 
 const PACKAGE = require("../package.json");
 const SCHEMA_VERSION = "v11.2.1";
-const SCHEMA_VERSIONS = [SCHEMA_VERSION, "v11.2.0", "v11.1.3", "v11.1.2", "v11.1.1"];
+const SCHEMA_VERSIONS = ["v11.2.1", "v11.2.0", "v11.1.3", "v11.1.2", "v11.1.1"];
+const ORDERED_SCHEMA_VERSIONS = ["v11.1.1", "v11.1.2", "v11.1.3", "v11.2.0", "v11.2.1"];
 
 const SUPPORTED_OPERATIONS = [
     "version", "describe",
@@ -13,7 +14,7 @@ const SUPPORTED_OPERATIONS = [
     "config"
 ];
 
-const SUPPORTED_EXTENSIONS = ["pre-request", "post-export", "pre-import", "multiline-text-diff", "policy-code-validator"];
+const SUPPORTED_EXTENSIONS = ["pre-request", "post-export", "pre-import", "post-revise", "multiline-text-diff", "policy-code-validator"];
 const SCHEMA_FEATURE_LIST = {
     "v11.2.1": ["mappings", "mappings-source", "policy-as-code"],
     "v11.2.0": ["mappings", "mappings-source", "policy-as-code"],
@@ -72,6 +73,7 @@ module.exports = {
         config.version = "v" + PACKAGE.version;
         config.defaultSchemaVersion = SCHEMA_VERSION;
         config.supportedSchemaVersions = SCHEMA_VERSIONS;
+        config.orderedSchemaVersions = ORDERED_SCHEMA_VERSIONS;
         config.supportedExtensions = SUPPORTED_EXTENSIONS;
         config.schemaVersion = String(params.options.schema || config.options.schema || SCHEMA_VERSION);
         config.schemaVersions = gqlschema.availableSchemas();
@@ -353,7 +355,7 @@ function makeOptions(options) {
         "policyCodeFormat": "xml",
         "keyFormat": "p12",
         "caFilename": null,
-        "extensions": ["pre-request", "post-export", "pre-import"]
+        "extensions": ["pre-request", "post-export", "pre-import", "post-revise"]
     }, options);
 }
 
@@ -377,7 +379,7 @@ function makeGateways(gateways) {
         gateways['default'] = {
             "address": "https://localhost:8443/graphman",
             "credential": "default",
-            "rejectUnauthorized": false,
+            "rejectUnauthorized": true,
             "passphrase": "7layer",
             "allowMutations": false
         };
